backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg==3.1.18
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
Pillow==12.0.0
django-sendgrid-v5==1.3.1
reportlab

